ai-engine/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
apscheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0
